+++ packages/shopify-codemod/test/fixtures/empty-func-to-lodash-noop/basic.output.js
@@ -1,0 +1,8 @@
+const foo = _.noop;
+foo(_.noop);
+foo(_.noop);
+
+const foo = function() { return 'foo'; };
+foo(function() { return 'bar'; });
+foo(() => { return 'bar'; });
+foo(() => 'bar');